+++ src/Button.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-const Button = ({ children, onClick }) => (
-  <button type="button" onClick={onClick}>
-    {children}
-  </button>
-);
-
-export default Button;
+export default function Button({ children, onClick }) {
+  return (
+    <button type="button" onClick={onClick}>
+      {children}
+    </button>
+  );
+}
