+++ src/Buttons.jsx
@@ -1,10 +1,11 @@
 import React from 'react';
+
 import Button from './Button';
 
-const Buttons = ({ onClick }) => (
-  [1, 2, 3, 4, 5].map((number) => (
-    <Button key={number} onClick={() => onClick(number)}>{number}</Button>
-  ))
-);
-
-export default Buttons;
+export default function Buttons({ onClick }) {
+  return (
+    [1, 2, 3, 4, 5].map((number) => (
+      <Button key={number} onClick={() => onClick(number)}>{number}</Button>
+    ))
+  );
+}
