+++ src/index.jsx
@@ -1,28 +1,6 @@
-import React, { useState } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
-import Counter from './Counter';
 
-const App = () => {
-  const [count, setCount] = useState(0);
-
-  function handleClickMe() {
-    setCount(count + 1);
-  }
-
-  function handleClickNumber(value) {
-    setCount(count + value);
-  }
-
-  return (
-    <div>
-      <p>Counter</p>
-      <Counter
-        count={count}
-        onClickMe={handleClickMe}
-        onClickNumber={handleClickNumber}
-      />
-    </div>
-  );
-};
+import App from './App';
 
 ReactDOM.render(<App />, document.querySelector('#app'));
